test(captions): relocate public-api test under _captions mirror (audit PS-204)
