+++ tests/figrecipe/_captions/test__convenience.py
@@ -18,3 +18,20 @@ def test_import__captions__convenience_module():
     mod = pytest.importorskip(module_path)
     # Assert
     assert mod.__name__ == module_path
+
+
+def test_caption_helpers_exposed_on_public_api():
+    """Card caption-public-api: add_figure_caption, add_panel_captions,
+    panel_label are reachable as figrecipe.X."""
+    # Arrange
+    import figrecipe as fr
+
+    # Act
+    exposed = [
+        getattr(fr, "add_figure_caption", None),
+        getattr(fr, "add_panel_captions", None),
+        getattr(fr, "panel_label", None),
+    ]
+
+    # Assert
+    assert all(callable(s) for s in exposed)
